+++ module-web/src/main/java/com/back2basics/post/controller/PostController.java
@@ -4,11 +4,11 @@ import com.back2basics.port.in.post.CreatePostUseCase;
 import com.back2basics.port.in.post.DeletePostUseCase;
 import com.back2basics.port.in.post.GetPostUseCase;
 import com.back2basics.port.in.post.UpdatePostUseCase;
+import com.back2basics.post.response.PostResponseCode;
 import com.back2basics.response.global.result.ApiResponse;
 import com.back2basics.service.post.dto.PostCreateCommand;
 import com.back2basics.service.post.dto.PostResponseDto;
 import com.back2basics.service.post.dto.PostUpdateCommand;
-import com.back2basics.service.post.response.PostResponseCode;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
